Add job queueing logic.
The JobQueue can queue jobs for multiple machines, with basic logic to avoid
queues blocking unecessarily (e.g. it doesn't block small-machine's queues if a
big-machine-only job is held up waiting for a big machine) and allow jobs to be
queued for many machines at once (and getting allocated only on the first which
becomes ready. The structure supports the dynamic addition, modification and
removal of machines and updates queues accordingly.