Controller improvements.
Move keepalive checking from a thread to a function called by the user.

Add sets which record which jobs/machines which have changed.

Add callback for when a power command completes in the background.

Move the Mahcine object outside and make it "properly" public and add careful
sanity checks to its constructor in preparation for this being a config file
interface.