Make allocating 'half' toruses produce no false-wrappings.
Previously if allocating a 'ring' of a machine (e.g. a region which wraps at
top and bottom but not left and right), the machine dimensions provided would
result in chips on the far right of the system being numbered as if on the far
left. This would mislead most routers/placers and generally be a bad thing. Now
in such situations, the dimensions reported should not experience this problem.